use GetAll to avoid multiple dbus call. Check slave interface exist before adding the bridge

+++ virt-bridge-setup.py
@@ -70,6 +70,7 @@ class NMManager:
             logging.error("Error connecting to D-Bus: %s", err)
             logging.error("Please ensure NetworkManager is running.")
             sys.exit(1)
+
     def select_default_slave_interface(self) -> Optional[str]:
         """
         Selects a default slave interface, prioritizing active devices with IP addresses.
@@ -87,18 +88,24 @@ class NMManager:
             for dev_path in devices_paths:
                 dev_proxy = self.bus.get_object('org.freedesktop.NetworkManager', dev_path)
                 prop_interface = dbus.Interface(dev_proxy, 'org.freedesktop.DBus.Properties')
-                iface = prop_interface.Get('org.freedesktop.NetworkManager.Device', 'Interface')
-                dev_type = prop_interface.Get('org.freedesktop.NetworkManager.Device', 'DeviceType')
+                all_props = prop_interface.GetAll('org.freedesktop.NetworkManager.Device')
 
-                if iface == 'lo' or dev_type == 5 or any(iface.startswith(p) for p in ['virbr', 'vnet', 'docker', 'p2p-dev-']):
+                iface = all_props['Interface']
+                dev_type = all_props['DeviceType']
+
+                if iface == 'lo' or dev_type == 5 or any(
+                        iface.startswith(p) for p in ['virbr', 'vnet', 'docker', 'p2p-dev-']):
                     continue
 
-                ip4_config_path = prop_interface.Get('org.freedesktop.NetworkManager.Device', 'Ip4Config')
+                ip4_config_path = all_props['Ip4Config']
                 has_ip = False
                 if ip4_config_path != "/":
-                    ip4_config_proxy = self.bus.get_object('org.freedesktop.NetworkManager', ip4_config_path)
-                    ip4_props_iface = dbus.Interface(ip4_config_proxy, 'org.freedesktop.DBus.Properties')
-                    if ip4_props_iface.GetAll('org.freedesktop.NetworkManager.IP4Config').get('Addresses'):
+                    ip4_config_proxy = self.bus.get_object(
+                            'org.freedesktop.NetworkManager', ip4_config_path)
+                    ip4_props_iface = dbus.Interface(
+                            ip4_config_proxy, 'org.freedesktop.DBus.Properties')
+                    if ip4_props_iface.GetAll(
+                            'org.freedesktop.NetworkManager.IP4Config').get('Addresses'):
                         has_ip = True
 
                 if dev_type == 1:  # Ethernet
@@ -113,7 +120,8 @@ class NMManager:
         for category in ['eth_with_ip', 'wifi_with_ip', 'eth_without_ip', 'wifi_without_ip']:
             if interface_lists[category]:
                 selected_iface = sorted(interface_lists[category])[0]
-                logging.info("Default slave interface selected: %s (%s)", selected_iface, category.replace('_', ' ').title())
+                logging.info("Default slave interface selected: %s (%s)",
+                             selected_iface, category.replace('_', ' ').title())
                 return selected_iface
 
         logging.warning("No suitable default slave interface was found.")
@@ -127,8 +135,10 @@ class NMManager:
             for dev_path in devices_paths:
                 dev_proxy = self.bus.get_object('org.freedesktop.NetworkManager', dev_path)
                 prop_interface = dbus.Interface(dev_proxy, 'org.freedesktop.DBus.Properties')
-                iface = prop_interface.Get('org.freedesktop.NetworkManager.Device', 'Interface')
-                dev_type = prop_interface.Get('org.freedesktop.NetworkManager.Device', 'DeviceType')
+                all_props = prop_interface.GetAll('org.freedesktop.NetworkManager.Device')
+
+                iface = all_props['Interface']
+                dev_type = all_props['DeviceType']
 
                 ignored_prefixes = ['lo', 'virbr', 'vnet', 'docker', 'p2p-dev-']
                 if dev_type not in [1, 2] or dev_type == 5 or any(
@@ -265,8 +275,11 @@ class NMManager:
         vlan_default_pvid = config.get('vlan_default_pvid', None)
         dry_run = config.get('dry_run', False)
 
-        slave_conn_name = f"{bridge_conn_name}-port-{slave_iface}"
+        if not self.check_interface_exist(slave_iface):
+            logging.error("Slave interface %s does not exist", slave_iface)
+            return
 
+        slave_conn_name = f"{bridge_conn_name}-port-{slave_iface}"
         self.delete_connection(bridge_conn_name, False, dry_run)
         self.delete_connection(slave_conn_name, False, dry_run)
 
@@ -372,11 +385,8 @@ class NMManager:
             dev_path = self.nm_interface.GetDeviceByIpIface(interface_name)
             dev_proxy = self.bus.get_object('org.freedesktop.NetworkManager', dev_path)
             prop_interface = dbus.Interface(dev_proxy, 'org.freedesktop.DBus.Properties')
-
-            ip4_config_path = prop_interface.Get(
-                                        'org.freedesktop.NetworkManager.Device',
-                                        'Ip4Config'
-                                        )
+            all_props = prop_interface.GetAll('org.freedesktop.NetworkManager.Device')
+            ip4_config_path = all_props['Ip4Config']
             if ip4_config_path == "/":
                 return None
 
@@ -628,18 +638,12 @@ class NMManager:
             for dev_path in devices_paths:
                 dev_proxy = self.bus.get_object('org.freedesktop.NetworkManager', dev_path)
                 prop_interface = dbus.Interface(dev_proxy, 'org.freedesktop.DBus.Properties')
-                #all_props = prop_interface.GetAll('org.freedesktop.NetworkManager.Device')
-                #dev_type_num = all_props['DeviceType']
-                iface = prop_interface.Get('org.freedesktop.NetworkManager.Device', 'Interface')
-                autoconnect_bool = prop_interface.Get(
-                                                    'org.freedesktop.NetworkManager.Device',
-                                                    'Autoconnect'
-                                                    )
+                all_props = prop_interface.GetAll('org.freedesktop.NetworkManager.Device')
+
+                iface = all_props['Interface']
+                autoconnect_bool = all_props['Autoconnect']
                 autoconnect_str = "Yes" if autoconnect_bool else "No"
-                dev_type_num = prop_interface.Get(
-                                                'org.freedesktop.NetworkManager.Device',
-                                                'DeviceType'
-                                                )
+                dev_type_num = all_props['DeviceType']
                 # WORKAROUND: Corrects known DeviceType bugs from certain NetworkManager versions.
                 if dev_type_num == 13 and ('br' in iface or 'virbr' in iface):
                     logging.debug(
@@ -653,15 +657,8 @@ class NMManager:
                         iface
                         )
                     dev_type_num = 28
-                dev_state_num = prop_interface.Get('org.freedesktop.NetworkManager.Device', 'State')
-                mac_address = "---"
-                try:
-                    mac_address = prop_interface.Get(
-                                                'org.freedesktop.NetworkManager.Device',
-                                                'HwAddress'
-                                                )
-                except dbus.exceptions.DBusException:
-                    pass
+                dev_state_num = all_props['State']
+                mac_address = all_props.get('HwAddress', '---')
 
                 active_conn_path = prop_interface.Get(
                                                 'org.freedesktop.NetworkManager.Device',
